pyproject.toml
[project]
name = "zonos"
version = "0.1.0"
description = "Text-to-speech by Zyphra"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.7.0",
    "setuptools",
    "packaging",
    "inflect>=7.5.0",
    "kanjize>=1.5.0",
    "numpy>=2.2.2",
    "phonemizer>=3.3.0",
    "sudachidict-full>=20241021",
    "sudachipy>=0.6.10",
    "torchaudio>=2.7.0",
    "transformers>=4.48.1",
    "soundfile>=0.13.1",
    "huggingface-hub>=0.28.1",
    "gradio>=5.15.0",
    "sounddevice>=0.5.3",
    "faster-whisper>=1.2.1",
    "torchcodec>=0.8.1",
    "whisper-timestamped>=1.15.9",
    "pydub>=0.25.1",
    "wavmark>=0.0.3",
    "eng-to-ipa>=0.0.2",
    "pypinyin>=0.55.0",
    "cn2an>=0.5.23",
    "jieba>=0.42.1",
    "langid>=1.1.6",
    "pygame>=2.6.1",
    "python-mecab-ko>=1.3.7",
    "mecab-python3>=1.0.10",
    "unidic-lite>=1.0.8",
    "fugashi>=1.5.2",
    "ipadic>=1.0.0",
    "pykakasi>=2.3.0",
    "g2pkk>=0.1.2",
    "jamo>=0.4.1",
    "anyascii>=0.3.3",
]



[tool.setuptools.packages.find]
include = ["zonos"]

[tool.uv]
no-build-isolation-package = ["flash-attn", "mamba-ssm", "causal-conv1d"]

[tool.uv.pip]
index-url = "https://download.pytorch.org/whl/cu118"
extra-index-url = ["https://pypi.org/simple"]

[tool.ruff]
line-length = 120
